+++ voting-booth-frontend/src/app/test.utilities.spec.ts
@@ -1,0 +1,31 @@
+/**
+ * Testing utilities
+ */
+/**
+ * Creates a mock store and spies on it
+ */
+export function establishLocalStorageSpies() {
+  let store = {};
+  const mockLocalStorage = {
+    getItem: (key: string): string => {
+      return key in store ? store[key] : null;
+    },
+    setItem: (key: string, value: string) => {
+      store[key] = `${value}`;
+    },
+    removeItem: (key: string) => {
+      delete store[key];
+    },
+    clear: () => {
+      store = {};
+    }
+  };
+  spyOn(localStorage, 'getItem')
+    .and.callFake(mockLocalStorage.getItem);
+  spyOn(localStorage, 'setItem')
+    .and.callFake(mockLocalStorage.setItem);
+  spyOn(localStorage, 'removeItem')
+    .and.callFake(mockLocalStorage.removeItem);
+  spyOn(localStorage, 'clear')
+    .and.callFake(mockLocalStorage.clear);
+}
